Renamed wrapper
